+++ example/basic_auth_server.py
@@ -18,4 +18,4 @@ async def main():
     ):
         await asyncio.Future()  # run forever
 
-asyncio.run(main)
+asyncio.run(main())