+++ example/echo.py
@@ -11,4 +11,4 @@ async def main():
     async with websockets.serve(echo, "localhost", 8765):
         await asyncio.Future()  # run forever
 
-asyncio.run(main)
+asyncio.run(main())