+++ example/health_check_server.py
@@ -20,4 +20,4 @@ async def main():
     ):
         await asyncio.Future()  # run forever
 
-asyncio.run(main)
+asyncio.run(main())